backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
PyMuPDF==1.24.3
python-docx==1.1.0
httpx==0.27.0
networkx==3.3
numpy==1.26.4
pydantic==2.7.1
reportlab==4.2.2
beautifulsoup4
